Adding OrderedBTreeDataFrame to keep track of the order that rows are added to a data frame. Also implementing this data frame for all OUSD playsheets.
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@9706 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 6e8816f38824e27acc09991fa44008e4efe35594

+++ src/prerna/ui/components/specific/ousd/ActivityGroupOptimizer.java
@@ -84,7 +84,7 @@ public class ActivityGroupOptimizer extends LPOptimizer{
 		 */
 		for(String system: sysList){
 			colno[i] = i + 1;
-			row[i] = treeMaxConstant * (double) (1-riskScoreMap.get(system)) * (double) sysBudgets.get(system);
+			row[i] = treeMaxConstant * (1-riskScoreMap.get(system)) * sysBudgets.get(system);
 			i++;
 		}
